requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.1
django-storages==1.9.1
django-url-filter==0.3.15